Compiler complaints handled
git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@703 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/net/general/net_city.cpp
@@ -63,7 +63,7 @@ NetCity::NetCity(UnitData* unit, BOOL isInitial) :
 	m_unitData(unit) 
 {
 	m_unitId = unit->m_id;
-	m_isInitialPacket = isInitial;
+	m_isInitialPacket = static_cast<uint8>(isInitial);
 }
 
 void NetCity::Packetize(uint8* buf, uint16& size)
